Shuffle utils into their relevant folders
