+++ src/mirage/resolvers/object.ts
@@ -1,10 +1,11 @@
 import { Resolver, ResolverInfo, ResolverParent } from '../../types';
-import { unwrap, coerceReturnType, coerceToList } from '../../utils/utils';
 import { MirageGraphQLMapper } from '../mapper';
 import { relayPaginateNodes } from '../../relay/helpers';
 import { mirageCursorForNode, ObjectResolverMatch, AutoResolverErrorMeta } from './helpers';
 import { extractDependencies } from '../../resolver/extract-dependencies';
 import { AutoResolverError } from './auto-resolver-error';
+import { unwrap } from '../../graphql/utils';
+import { coerceToList, coerceReturnType } from '../../resolver/utils';
 
 function findMatchingFieldForObjectParent({
   mirageMapper,
