+++ src/mirage/resolvers/root-query.ts
@@ -1,7 +1,6 @@
 import { Server as MirageServer } from 'miragejs';
 import { isScalarType } from 'graphql';
 import { Resolver } from '../../types';
-import { unwrap, coerceReturnType, coerceToList } from '../../utils/utils';
 import { MirageGraphQLMapper } from '../mapper';
 import { relayPaginateNodes } from '../../relay/helpers';
 import {
@@ -12,6 +11,8 @@ import {
 } from './helpers';
 import { extractDependencies } from '../../resolver/extract-dependencies';
 import { AutoResolverError } from './auto-resolver-error';
+import { unwrap } from '../../graphql/utils';
+import { coerceToList, coerceReturnType } from '../../resolver/utils';
 
 function findMatchingModelsForType({
   type,
